+++ 42/media/lua/shared/PhobosLib.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------
 
 PhobosLib = PhobosLib or {}
-PhobosLib.VERSION = "1.0.0"
+PhobosLib.VERSION = "1.1.0"
 
 require "PhobosLib_Util"
 require "PhobosLib_Fluid"
@@ -13,3 +13,4 @@ require "PhobosLib_World"
 require "PhobosLib_Sandbox"
 require "PhobosLib_Quality"
 require "PhobosLib_Hazard"
+require "PhobosLib_Skill"